Refactor dashboard components and contexts
- Removed MainBoard, ProfileSettings, ResetPassword, BoardContext, and SubdomainContext components.
- Introduced new BoardGroup, CardItem, ListItem, and TableView components for better modularity.
- Updated Dashboard to utilize new component structure.
- Enhanced sorting and filtering functionality in BoardGroup.
- Improved loading states and error handling across components.
