+++ src/features/dashboard/components/TableView.jsx
@@ -1,6 +1,6 @@
 import { useNavigate } from "react-router-dom";
 import PropTypes from "prop-types";
-import useUsersPhotoThumbs from "../../hooks/useUsersPhotoThumbs";
+import useUsersPhotoThumbs from "../../../hooks/useUsersPhotoThumbs";
 import { Loader } from "@mantine/core";
 
 const TableView = ({ item }) => {
